Organized top-level schematic.
Mapped LEDs[0..2] to ps2_msg[0..2] and LEDs[8..10] to ps2_msg[8..10].
Tested that these are correct outputs using oscope.

Found that LEDs which are LOW (i.e. 0) in the ps2_msg would be dimly
lit when the button on the keyboard was continuously held pressed.
I think this is because the shift register is continuous meaning on
each pulse of data_sample_clk all bits in ps2_msg are shifted;  this
means that the 0 bits in the ps2_msg will have 1s propagated through
them.  While 1s are being propogated through them, they light up.