Keycodes displayed on leds
MAKE SURE YOU RESET BEFORE RUNNING!  This commit can parse the passed
ps2 data from the keyboard into the respective keycodes.  Each key
transmits a "makecode" when pressed, and a "break code" when released.
The "break code" is ussually two codes back-to-back.  A release code
followed by the makecode.  See the links in docs/ps2_hardware_notes.txt
for more info.

The last couple of steps to finish ps2_bring_up are the following:
create a button driver so that keyboard is like a collection of buttons
where each button has its own signal.  Encapsulate the ps2_bring_up
project into a driver that takes a couple of inputs and produces a
couple of outputs.